# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AlgDevs-AI Agent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Plans ─────────────────────────────────────────────────────────────────
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a new plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get a plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanWithSteps"
        "404":
          description: Not found
    patch:
      operationId: updatePlan
      tags: [plans]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /plans/{id}/execute:
    post:
      operationId: executePlan
      tags: [plans]
      summary: Start executing a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}/steps/{stepId}:
    patch:
      operationId: updatePlanStep
      tags: [plans]
      summary: Update a plan step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanStepUpdate"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanStep"

  # ── Tasks ─────────────────────────────────────────────────────────────────
...
  /artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      summary: List all artifacts
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: ["integer", "null"]
        - name: artifactType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artifact"
    post:
      operationId: createArtifactRecord
      tags: [artifacts]
      summary: Create an artifact record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtifactInput"
      responses:
        "201":
          description: Created artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
  /artifacts/{id}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: Get an artifact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "404":
          description: Not found
    delete:
      operationId: deleteArtifact
      tags: [artifacts]
      summary: Delete an artifact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Workspaces ────────────────────────────────────────────────────────────
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List active agent instances
      responses:
        "200":
          description: List of agent instances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentInstance"
    post:
      operationId: spawnAgent
      tags: [agents]
      summary: Spawn a new agent instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSpawnInput"
      responses:
        "201":
          description: Spawned agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInstance"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent instance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInstance"
  /agents/{id}/stop:
    post:
      operationId: stopAgent
      tags: [agents]
      summary: Stop an agent instance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInstance"

  # ── Sandbox ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        agentType:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        tokensUsed:
          type: ["integer", "null"]
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        model:
          type: ["string", "null"]
        providerId:
          type: ["integer", "null"]

    # ── Plan schemas ─────────────────────────────────────────────────────────
    Plan:
      type: object
      required: [id, title, status, totalSteps, completedSteps, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, executing, completed, failed, paused]
        totalSteps:
          type: integer
        completedSteps:
          type: integer
        sessionId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
        version:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    PlanStep:
      type: object
      required: [id, planId, title, order, status, createdAt]
      properties:
        id:
          type: integer
        planId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        agentType:
          type: ["string", "null"]
        output:
          type: ["string", "null"]
        dependsOn:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PlanWithSteps:
      type: object
      required: [id, title, status, totalSteps, completedSteps, steps, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, executing, completed, failed, paused]
        totalSteps:
          type: integer
        completedSteps:
          type: integer
        sessionId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
        version:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PlanStep"
        createdAt:
          type: string
        updatedAt:
          type: string
    PlanInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        sessionId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
        steps:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
              description:
                type: string
              order:
                type: integer
              agentType:
                type: string
              dependsOn:
                type: array
                items:
                  type: integer
    PlanUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, executing, completed, failed, paused]
    PlanStepUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        output:
          type: ["string", "null"]

    # ── Task schemas ─────────────────────────────────────────────────────────
...
    Artifact:
      type: object
      required: [id, title, artifactType, creatorAgent, version, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artifactType:
          type: string
          enum: [markdown, code, json, csv, report, zip, text, other]
        content:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        creatorAgent:
          type: string
        taskId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]
        version:
          type: integer
        sizeBytes:
          type: ["integer", "null"]
        createdAt:
          type: string
    ArtifactInput:
      type: object
      required: [title, artifactType, creatorAgent]
      properties:
        title:
          type: string
          minLength: 1
        artifactType:
          type: string
          enum: [markdown, code, json, csv, report, zip, text, other]
        content:
          type: string
        filePath:
          type: ["string", "null"]
        creatorAgent:
          type: string
        taskId:
          type: ["integer", "null"]
        workspaceId:
          type: ["integer", "null"]

    # ── Workspace schemas ────────────────────────────────────────────────────
...
    AgentInstance:
      type: object
      required: [id, agentType, status, createdAt]
      properties:
        id:
          type: integer
        agentType:
          type: string
          enum: [supervisor, planner, researcher, coder, reviewer]
        status:
          type: string
          enum: [idle, running, paused, stopped, error]
        currentTaskId:
          type: ["integer", "null"]
        sessionId:
          type: ["integer", "null"]
        model:
          type: ["string", "null"]
        tokensUsed:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    AgentSpawnInput:
      type: object
      required: [agentType]
      properties:
        agentType:
          type: string
          enum: [supervisor, planner, researcher, coder, reviewer]
        sessionId:
          type: ["integer", "null"]
        model:
          type: ["string", "null"]

    # ── Sandbox schemas ──────────────────────────────────────────────────────
...