# lib/api-zod/src/generated/types/artifactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlgDevs-AI Agent Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ArtifactInputArtifactType } from './artifactInputArtifactType';

export interface ArtifactInput {
  /** @minLength 1 */
  title: string;
  artifactType: ArtifactInputArtifactType;
  content?: string;
  /** @nullable */
  filePath?: string | null;
  creatorAgent: string;
  /** @nullable */
  taskId?: number | null;
  /** @nullable */
  workspaceId?: number | null;
}