# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlgDevs-AI Agent Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanStatus } from './planStatus';

export interface Plan {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: PlanStatus;
  totalSteps: number;
  completedSteps: number;
  /** @nullable */
  sessionId?: number | null;
  /** @nullable */
  workspaceId?: number | null;
  version?: number;
  createdAt: string;
  updatedAt: string;
}