# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlgDevs-AI Agent Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanUpdateStatus } from './planUpdateStatus';

export interface PlanUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: PlanUpdateStatus;
}